be/requirement.txt
Flask==2.2.3
Flask-SQLAlchemy==3.1.1
flask-cors==4.0.0
mysql-connector-python==8.1.0
pandas==2.1.1
joblib==1.3.2
scikit-learn==1.3.1
python-dotenv==1.0.0
mysqlclient==2.2.0